Add root command in Main file
This root command will be the main entry point for the client and the
server
