+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ch.heigvd</groupId>
-    <artifactId>connect-4</artifactId>
+    <artifactId>connect4</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -26,35 +26,6 @@
     <build>
         <plugins>
             <plugin>
-                <groupId>com.diffplug.spotless</groupId>
-                <artifactId>spotless-maven-plugin</artifactId>
-                <version>2.46.1</version>
-                <configuration>
-                    <pom>
-                        <sortPom>
-                            <expandEmptyElements>false</expandEmptyElements>
-                            <sortPlugins>groupId,artifactId</sortPlugins>
-                        </sortPom>
-                    </pom>
-                    <java>
-                        <googleJavaFormat>
-                            <formatJavadoc>true</formatJavadoc>
-                        </googleJavaFormat>
-                        <importOrder/>
-                        <removeUnusedImports/>
-                        <formatAnnotations/>
-                    </java>
-                </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>check</goal>
-                        </goals>
-                        <phase>package</phase>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
                 <version>3.6.0</version>
@@ -67,7 +38,7 @@
                         <configuration>
                             <transformers>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>ch.heigvd.Main</mainClass>
+                                    <mainClass>ch.heigvd.connect4.Main</mainClass>
                                 </transformer>
                                 <transformer implementation="org.apache.maven.plugins.shade.resource.DontIncludeResourceTransformer">
                                     <resource>MANIFEST.MF</resource>
