+++ src/main/java/ch/heigvd/connect4/utils/ClientCommands.java
@@ -1,4 +1,4 @@
-package connect4;
+package ch.heigvd.connect4.utils;
 
 /**
  * Defines the command the client can send
